+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.25.0
 - Adapter API: Swagger
 - Adapter API and events: add static string property (req for tmForum APIs)
+- Inherit entity from existing one 
 
 # 0.24.4
 - state: hide create dialog if apiManaged